pyproject.toml
# pyproject.toml
[project]
name = "ai-git"
version = "0.2.0"
description = "Generate Git commit messages using Large Language Models (LLM) like GPT and Qwen through OpenRouter"
dependencies = [
    "openai>=1.0.0",
    "typer>=0.6.1",
    "requests",
    "rich>=13.0.0",
    "inquirer>=3.1.3",
    "pyyaml>=6.0.2",
]

[project.scripts]
ai-git = "ai_git.cli:app"

[tool.setuptools]
packages = ["ai_git"]
